Fix viscosity force sign and GPU ParticleCount upload
The viscosity force used the spiky kernel derivative directly, which is
negative inside the smoothing radius, so positive ViscosityFactor values
amplified velocity differences instead of damping them. Negate the kernel
term so viscosity pulls neighboring velocities together.

The compute shader declares ParticleCount as int, but it was uploaded with
SetFloat, writing raw float bits into the constant buffer. Since
ParticleCount is the spatial hash modulus and neighbor loop bound, the GPU
neighbor search was effectively broken. Use SetInt.

+++ Assets/fluid-sim/GPU/FluidSimGPU.cs
@@ -164,7 +164,7 @@ public class FluidSimGPU : MonoBehaviour, IFluidSim
             externalForceStrength *= -1;
         }
         
-        SimComputeShader.SetFloat("ParticleCount", ParticleCount);
+        SimComputeShader.SetInt("ParticleCount", ParticleCount);
         SimComputeShader.SetFloat("Mass", Mass);
         SimComputeShader.SetFloat("Width", Width);
         SimComputeShader.SetFloat("Height", Height);
